requirements.txt
pathlib==1.0.1
pillow==11.3.0
